+++ test/api_drivers.test.ts
@@ -170,6 +170,22 @@ describe('gdal.drivers', () => {
       assert.equal(ds.driver.description, 'MEM')
     })
 
+    it('should support progress callbacks', () => {
+      const driver = gdal.drivers.get('MEM')
+      const outputFilename = ''
+      let calls = 0
+      const ds = driver.createCopy(
+        outputFilename,
+        gdal.open(`${__dirname}/data/12_791_1476.jpg`),
+        {},
+        false,
+        (): void => {
+          calls++
+        })
+      assert.equal(ds.driver.description, 'MEM')
+      assert.isAbove(calls, 0)
+    })
+
     it('should throw on error', () => {
       const driver = gdal.drivers.get('GTiff')
       const outputFilename = '' // __dirname + '/data/12_791_1476.tif';